Modifications to the main mock simulation
In this update, we go from a one car system to a two car system. Unfortunately, due to the time left before our deadline, we won't be able to incorporate an algorithm for for than 2 cars. The second car and it's nature was hard coded in. Future progression of this code would be to maybe incorporate a way to use threads to create parallelism between functions and possibly a way to easily create car objects